report-management - rename to _context
