+++ tests/bark/add-appointment.test.ts
@@ -1,5 +1,5 @@
 import { CODE } from "@/lib/utilities/bark-code";
-import { withBarkContext } from "./_service";
+import { withBarkContext } from "./_context";
 import { givenDog, givenVet } from "./_given";
 import { BarkAction_addAppointment } from "@/lib/bark/actions/add-appointment";
