+++ tests/bark/get-report.test.ts
@@ -1,7 +1,7 @@
 import { BarkReport, BarkReportData } from "@/lib/bark/bark-models";
 import { givenDog, givenVet } from "./_given";
 import { mockReportData } from "./_mocks";
-import { withBarkContext } from "./_service";
+import { withBarkContext } from "./_context";
 import { BarkAction_addAppointment } from "@/lib/bark/actions/add-appointment";
 import { BarkAction_createReport } from "@/lib/bark/actions/create-report";
 import { BarkAction_getReport } from "@/lib/bark/actions/get-report";
